requirements.txt
pycryptodome 
dill
PyYAML
streamlit_authenticator
PyAudio
SpeechRecognition
